refactor: move decodeMerklClaimAmount function to mUSD utility and clean up constants
