+++ app/components/UI/Earn/components/MerklRewards/constants.ts
@@ -1,5 +1,4 @@
 import { Hex } from '@metamask/utils';
-import { Interface } from '@ethersproject/abi';
 
 export const MERKL_API_BASE_URL = 'https://api.merkl.xyz/v4';
 
@@ -28,31 +27,3 @@ export const DISTRIBUTOR_CLAIMED_ABI = [
 export const DISTRIBUTOR_CLAIM_ABI = [
   'function claim(address[] calldata users, address[] calldata tokens, uint256[] calldata amounts, bytes32[][] calldata proofs)',
 ];
-
-/**
- * Decode the claim amount from a Merkl claim transaction data.
- * The claim function signature is: claim(address[] users, address[] tokens, uint256[] amounts, bytes32[][] proofs)
- *
- * @param data - The transaction data hex string
- * @returns The first claim amount as a string (raw value, not adjusted for decimals), or null if decoding fails
- */
-export function decodeMerklClaimAmount(
-  data: string | undefined,
-): string | null {
-  if (!data || typeof data !== 'string') {
-    return null;
-  }
-
-  try {
-    const contractInterface = new Interface(DISTRIBUTOR_CLAIM_ABI);
-    const decoded = contractInterface.decodeFunctionData('claim', data);
-    // amounts is the 3rd parameter (index 2)
-    const amounts = decoded[2];
-    if (!amounts || amounts.length === 0) {
-      return null;
-    }
-    return amounts[0].toString();
-  } catch {
-    return null;
-  }
-}
