+++ app/components/UI/TransactionElement/utils.js
@@ -35,8 +35,10 @@ import { calculateTotalGas, renderGwei } from './utils-gas';
 import { getTokenTransferData } from '../../Views/confirmations/utils/transaction-pay';
 import { hasTransactionType } from '../../Views/confirmations/utils/transaction';
 import { BigNumber } from 'bignumber.js';
-import { decodeMerklClaimAmount } from '../Earn/components/MerklRewards/constants';
-import { convertMusdClaimAmount } from '../Earn/utils/musd';
+import {
+  convertMusdClaimAmount,
+  decodeMerklClaimAmount,
+} from '../Earn/utils/musd';
 
 const POSITIVE_TRANSFER_TRANSACTION_TYPES = [
   TransactionType.perpsDeposit,
